+++ src/Estimators/src/UkfMeasurement.cpp
@@ -136,6 +136,11 @@ bool UkfMeasurement::finalize(const System::VariablesHandler& handler)
     return true;
 }
 
+void UkfMeasurement::setStateNameMapping(const std::map<std::string, std::string>& stateToUkfNames)
+{
+    m_stateToUkfNames = stateToUkfNames;
+}
+
 std::unique_ptr<UkfMeasurement>
 UkfMeasurement::build(std::weak_ptr<const ParametersHandler::IParametersHandler> handler,
                       System::VariablesHandler& stateVariableHandler,
@@ -209,13 +214,14 @@ UkfMeasurement::build(std::weak_ptr<const ParametersHandler::IParametersHandler>
         dynamicsGroup->setParameter("sampling_time", measurement->m_dT);
 
         // create variable handler
-        std::string dynamicsName;
-        std::vector<double> covariances;
-        if (!dynamicsGroup->getParameter("name", dynamicsName))
+        std::string inputName;
+        if (!dynamicsGroup->getParameter("input_name", inputName))
         {
-            BipedalLocomotion::log()->error("{} Unable to find the parameter 'name'.", logPrefix);
+            BipedalLocomotion::log()->error("{} Unable to find the parameter 'input_name'.",
+                                            logPrefix);
             return nullptr;
         }
+        std::vector<double> covariances;
         if (!dynamicsGroup->getParameter("covariance", covariances))
         {
             BipedalLocomotion::log()->error("{} Unable to find the parameter 'covariance'.",
@@ -238,16 +244,18 @@ UkfMeasurement::build(std::weak_ptr<const ParametersHandler::IParametersHandler>
                                             "the measurement variable `{}`.",
                                             logPrefix,
                                             dynamicModel,
-                                            dynamicsName);
+                                            dynamicsGroupName);
             return nullptr;
         }
 
         dynamicsInstance->setSubModels(subModelList, kinDynWrapperList);
 
-        dynamicsInstance->initialize(dynamicsGroup);
+        dynamicsInstance->initialize(dynamicsGroup, dynamicsGroupName);
 
         // add dynamics to the list
-        measurement->m_dynamicsList.emplace_back(dynamicsName, dynamicsInstance);
+        measurement->m_dynamicsList.emplace_back(dynamicsGroupName, dynamicsInstance);
+
+        measurement->m_ukfNamesToMeasures[dynamicsGroupName] = inputName;
     }
 
     measurement->m_inputDescription
@@ -324,7 +332,7 @@ UkfMeasurement::predictedMeasure(const Eigen::Ref<const Eigen::MatrixXd>& curren
         const_cast<UkfMeasurement*>(this)->m_ukfInput.robotJointAccelerations
             = m_jointAccelerationState;
 
-        for (int indexDyn = 0; indexDyn < m_dynamicsList.size(); indexDyn++)
+       for (int indexDyn = 0; indexDyn < m_dynamicsList.size(); indexDyn++)
         {
             m_dynamicsList[indexDyn].second->setState(m_currentState);
 
@@ -394,31 +402,28 @@ bool UkfMeasurement::freeze(const bfl::Data& data)
 
     m_measurementMap = bfl::any::any_cast<std::map<std::string, Eigen::VectorXd>>(data);
 
-    for (int index = 0; index < m_dynamicsList.size(); index++)
+    for (const auto& [key, value] : m_ukfNamesToMeasures)
     {
-        m_offsetMeasurement
-            = m_measurementVariableHandler.getVariable(m_dynamicsList[index].first).offset;
+        m_offsetMeasurement = m_measurementVariableHandler.getVariable(key).offset;
 
-        if (m_measurementMap.count(m_dynamicsList[index].first) == 0)
+        if (m_measurementMap.count(key) == 0)
         {
             BipedalLocomotion::log()->error("{} Measurement with name `{}` not found.",
                                             logPrefix,
-                                            m_dynamicsList[index].first);
+                                            key);
             return false;
         }
 
         // If more sub-models share the same accelerometer or gyroscope sensor, the measurement
         // vector is concatenated a number of times equal to the number of sub-models using the
         // sensor.
-        while (m_offsetMeasurement
-               < (m_measurementVariableHandler.getVariable(m_dynamicsList[index].first).offset
-                  + m_measurementVariableHandler.getVariable(m_dynamicsList[index].first).size))
+        while (m_offsetMeasurement < (m_measurementVariableHandler.getVariable(key).offset
+                                      + m_measurementVariableHandler.getVariable(key).size))
         {
-            m_measurement.segment(m_offsetMeasurement,
-                                  m_measurementMap[m_dynamicsList[index].first].size())
-                = m_measurementMap[m_dynamicsList[index].first];
+            m_measurement.segment(m_offsetMeasurement, m_measurementMap[key].size())
+                = m_measurementMap[key];
 
-            m_offsetMeasurement += m_measurementMap[m_dynamicsList[index].first].size();
+            m_offsetMeasurement += m_measurementMap[key].size();
         }
     }